13.06.2023
1) Changed DaoException class extend from RuntimeException instead from
Exception.

2) Added check situation in findAllLessOrEqual method :
if the result is empty, so no one group this student's count
it returns a message to the user.

3) Added check situation in findAllLessOrEqual method :
Trying to get numeric value until user don't enter it.

4) Re-named getInfo method to getStudentById in DaoStudent
5) Re-named reader class to "Util"
6) Added access modifier to Course Service Impl Class
7) Deleted unnecessary part of code from StudentServiceImpl (getById
method)
8) Added the deleted part to command logic
9) Overrided ToString method in student entity to print or not the
course list - so if i use ToString method for a Student, if this object
has an empty Courses list it doesn't print at all the text "Course List:
" with the empty spaces as result. Could be useful if i need just
Student personal Information.
10) Added check situation for Delete Student Command & Add Course to
Student - if a student does not exist with entered ID - inform user
about it.
11) Added check for user input when adding student, input should be
alphabetical characters.
12) Deleted sc.close in every command method, so now don't have the
Empty line reading error that i fall in for entire last week.
13) Added 3 two method for dao and services (setGroupById & getAll &
getStudentByName) to set the group to the new student.

+++ schoolconsoleapp/src/main/java/ua/com/foxminded/yuriy/schoolconsoleapp/commands/StudentCommandsImpl/DeleteStudentCommand.java
@@ -16,16 +16,31 @@ public class DeleteStudentCommand implements Command {
 	public void execute() throws DaoException {
 		System.out.println("Enter student's ID you want to delete..");
 		Scanner sc = new Scanner(System.in);
-		int id = sc.nextInt();
-		System.out.println(studentService.getInfo(id));
-		System.out.println("Enter - 1 to confirm and - 2 to cancel.");
-		int confirmation = sc.nextInt();
-		if (confirmation == 1) {
-			studentService.delete(id);
-		} else {
-			System.out.println("You canceled the operation.");
+
+		while (!sc.hasNextInt()) {
+			sc.next();
+			System.out.println("You should enter a numeric value, please retry.");
 		}
-		sc.close();
+
+		int id = sc.nextInt();
+		Student student = studentService.getById(id);
+
+		if (student.getFirstName() == null) {
+			System.out.println("Student with following ID : " + "[ " + id + " ] is not found.");
+		} else {
+			System.out.println("Are you sure that you want to delete: " + studentService.getById(id).toString());
+			System.out.println("Enter - 1 to confirm and - 2 to cancel.");
+			while (!sc.hasNextInt()) {
+				sc.next();
+				System.out.println("You should enter a numeric value, please retry.");
+			}
+			int confirmation = sc.nextInt();
+			if (confirmation == 1) {
+				studentService.delete(id);
+			} else {
+				System.out.println("You canceled the operation.");
+			}
+		}
 	}
 
 	@Override
